add prettier
