+++ frontend/src/pages/index.tsx
@@ -1,0 +1,21 @@
+import type { NextPage } from 'next';
+import Head from 'next/head';
+import Image from 'next/image';
+import styles from '../styles/Home.module.css';
+
+const Home: NextPage = () => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>monoswap</title>
+        <meta name='description' content='' />
+        <link rel='icon' href='/favicon.ico' />
+      </Head>
+      <main>
+        <h1 className='text-3xl font-bold underline'>Hello world!</h1>
+      </main>
+    </div>
+  );
+};
+
+export default Home;
